Remove custom element upgrades.
This removes the upgrade candidate machinery so now if an
element is created before it's registered it'll never get
turned into the correct type. created() callbacks and the
correct wrappers being created can stil happen async though
because certain operations are not safe to run script inside
of (ex. the parser or editing).

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/843063005