Eliminate Mac Android Debug Engine shard (#27980)
At one point, this shard built the Android Vulkan binaries, but during
the migration to ci.yaml-based recipes, the gn and build steps for that
build were accidentally removed. As such, this shard no longer builds
anything.

This was detected when a buildroot roll triggered a clobber build which
caused the verify_exported.dart script to fail due to a missing 'out'
directory, which caused several of us to wonder why we were:
1. Checking that we strip release binaries on a debug bot, and
2. Why we're checking that we strip release binaries before we actually
   build anything at all.

It appears that we used to build android vulkan binaries, such as in
steps 14-16 of Build 11130:
* https://ci.chromium.org/ui/p/flutter/builders/prod/Mac%20Android%20Debug%20Engine/11130/overview

We migrated to ci.yaml-based infra in this commit:
* https://flutter-review.googlesource.com/c/infra/+/15660

A no-op rebuild was triggered in commmit:
* https://github.com/flutter/engine/pull/27599

Which triggered the following buikld, Build 11131:
* https://ci.chromium.org/ui/p/flutter/builders/prod/Mac%20Android%20Debug%20Engine/11131/overview

That build does not contain the gn or build steps for the Android Vulkan
binaries.

HOWEVER... we have no users actually using Android Vulkan binaries, so
rather than update the build to fix this, we delete the shard altogether
and save some CPU.

This will be followed by a patch that removes the --enable-vulkan flag
from gn and throughout our build.