[Impeller] Start of Dart GPU prototype (#42228)
I've been slowly hacking on a prototype alongside a doc (public go link [imminent](https://github.com/flutter/website/pull/8716)) for a while (the doc has been changing a lot as a result of prototyping) and I think it's time to start landing parts of this prototype/experiment in-tree.

This initial PR just sets up the main context singleton on the UI thread. After this, I'll land the shader management stuff.

I re-used the existing experimental 3D flag for this, since this is meant to replace what can be done with Impeller Scene anyway.